mpir_pmi: add auto-detection of PMI version 1,2,x at runtime.
Add new default value for MPIR_CVAR_PMI_VERSION=auto to auto-detect
if a process is running in a PMI1, PMI2 or PMIx environment.

The env var PMI_SIZE is used for auto-detection of PMI1 and PMI2.
If PMI_SIZE is not found PMI_PORT and PMI_ID are checked to cover
hydra's -pmi-port option.

The env var PMIX_NAMESPACE is used to auto-detect PMIx.

If both interfaces are detected a respective error is returned and
startup is aborted. If a PMI version is detected at runtime which
is not supported by the MPI library an error is returned and startup
is aborted.