mpl: Improve local host detection
Lookup the IP address of the hostname when determining if it is the
local host. Fixes an issue where the hostlist provided by the user or
resource manager does not contain an exact match for the hostname
returned by gethostname. For example, on LCRC Swing the PBS hostlist
contains FQDNs like "gpu3.lcrc.anl.gov", while gethostname returns only
"gpu3". Fixes pmodels/mpich#7186.