pmi/v1: close stdin before exec mpiexec in singleton
When singleton launches mpiexec, it is unclear which process should take
stdin. Since normally mpiexec (or its proxies) do not take stdin and
will forward stdin to the first process anyway, let's simply close stdin
in the singleton case to avoid confusion.

This fixes `run_mpitests` in the singleton tests. run_mpitests
communicates with the mpi process via stdin, thus it is crucial for
mpiexec not to grab the stdin away.