test/pt2pt: Remove truncated message from rqfree test
Posting a too small receive buffer is erroneous and makes testing
otherwise correct functionality difficult. For example, if MPI does not
catch the truncation error, then the system runtime may crash when
writing past the end of the receive buffer. We have truncation specific
tests which we can rely on instead.