pm/hydra: Pass proxy structs by reference to launch functions
During a spawn operation, we pass a "filtered" proxy list to the launch
function. For bookkeeping purposes we map local pids to their proxy
structures. For spawn this was broken because we stored pointers to the
filtered list which is freed directly after launch. To fix, we pass
references to the canonical proxy structures stored in the pg and avoid
potential use-after-free issues. Fixes pmodels/mpich#7528.