coll/gen_coll: dump algo functions
Dump a wrapper function for each algorithm that takes (cont, coll_sig).

Separately Declare algorithm prototypes. Note that the old code in
dump_coll() already declared algorithm prototypes, but it will be
removed later. It is okay to temporarily declare twice as long as the
declarations are consistent.

Separately Decleare sched_auto prototypes.