all: use fallback routines for internal collectives
The fallback collectives (e.g. MPIR_Bcast_fallback) are manual "auto"
functions that may not be the best algorithms for the system, but
are sufficient for internal usages during init and object
constricutions.

Not all collective types have fallbacks defined since internally we use
limited types. We'll define new fallback routines when we need them in
the future.

This prepares for the revamp of CSEL.