coll/csel: update csel
Generate tables based on coll_algorithms.txt and use the tables to
facilitate csel parsing and error reporting.

If user sets an algorithm CVAR, directly construct a container for the
cvar-specified algorithm and call it if all restrictions are met.

All restrictive checkers are represented by either a bit in
coll_sig->attr or a boolean checker function. All restrictions and their
checkers are configured in coll_algorithms.txt.