test/config: Error out if Fortran requested but cannot be compiled
The testsuite tries to compile a simple program with each of mpif.h, use
mpi, and use mpi_f08 and will disable tests for any unsuccessful
compile. It can be desirable, e.g. in CI builds, to force support for
Fortran tests and thus return an error if programs cannot be
compiled. By default, we still auto-detect the supported Fortran
methods, but now passing --enable-fortran will cause configure to error
if any of the three do not work. Fixes pmodels/mpich#7559.