ucc: add support for UCC in mpid/common
This adds cross-device support for Unified Collective Communication
(UCC) to `mpid/common` alongside the existing hcoll support.

The bindings design is broadly based on that of hcoll, but the
implementation is also partly inspired by that found in Open MPI.

Currently, only the following *blocking* collectives are supported:
 - MPI_Barrier
 - MPI_Bcast
 - MPI_Gather
 - MPI_Gatherv
 - MPI_Scatter
 - MPI_Scatterv
 - MPI_Reduce
 - MPI_Reduce_scatter
 - MPI_Reduce_scatter_block
 - MPI_Allgather
 - MPI_Allgatherv
 - MPI_Allreduce
 - MPI_Alltoall
 - MPI_Alltoallv

However, support for non-blocking operations should be quite straight-
forward to add.