mpi_t: Rename variable for clarity
Whether or not the user has set an environment variable is tracked by
the return codes of various MPL_env2* functions. If a variable is set
and debugging is enabled, the setting gets printed out for the
user. Rename got_rc->is_set to better indicate what the value
represents.