config: Check for Python 3 if user sets PYTHON
We were skipping the check for Python 3 if the configuration environment
had a PYTHON variable set. If this variable pointed to something other
than a Python 3 interpreter, it would result in broken Fortran bindings
that could not be built. Add a sanity check to make sure PYTHON works
and return an error if not. Fixes pmodels/mpich#7576.