autogen: patch libtool for nagfor
When nagfor is used as linker, it will pass to gcc, and then gcc will
pass to ld. Thus, ld options are passed with -Wl,-Wl,, prefix. The ld
options to build so library from convenience library is -

    --whole-archive convenience-libs --no-whole-archive

However, due to the double-pass, nagfor need pass options like -

    -Wl,-Wl,,--whole-archive -Wl,convenience-libs -Wl,-Wl,,--no-whole-archive

In particular, the list of convenience-libs need be prefixed with -Wl,
in order to be passed to gcc in the correct order.