coll: add MPIR_CVAR_DUMP_COLL_ALGO_COUNTERS
Provide a simple mechanism for a rank to dump collective algorithm
counters.

Set MPIR_CVAR_DUMP_COLL_ALGO_COUNTERS to the global rank of the process
that we want it to dump since it is undesirable for every process to
dump yet it does not always makes sense for rank 0 to dump especially
when we don't always use comm world.

It is counted in the CSEL framework so internal collectives are not
counted when we internally use _fallback algorithms.