datatype: re-add C types to `MPIR_Type_match_size_impl`
The C datatypes were removed in c7fc51b as only the hardcoded constants
in `mpi.h` were necessary to check for size matches. However, with
internal datatype checking, the C types must be re-appended so they can
be matched even if Fortran types are not present (i.e. with
`--disable-fortran`).