ch4/ofi: ensure send chunks in order in ofi_pipeline
If we make individual chunks as separate async tasks, they may be
completed out-of-order and consequently be sent out-of-order, resulting
mismatch to receive chunks. The out-of-order may happen as the chunk
completion may take place in between the progress iterating async task
list.

The solution is to combine all chunk copies as a single async task
and maintain the outstanding chunk list inside the task. To ensure
order, only check the completion at the head of the list.