typerep: only calculate size with `typeptr` when not a built-in
`typeptr` is incorrectly used by `MPIR_Typerep_size_external32` when
calculating the size of built-in datatypes. If the size of a built-in
and the size of its internal type pointed by `typeptr` differ, an
incorrect size is calculated. Moving the layout size calculation to only
happen when `type` is not a built-in fixes this behavior.