coll: check MPIR_CVAR_COLLECTIVE_FALLBACK
When user use a CVAR to selelct an algorithm, if the algorithm is a
compositional algorithm, we may run into recursive dead-loop. The
following measures breaks the cycle:

* Proper compositional algorithm should specify complete restriction
check that fails recursively. The default MPIR_CVAR_COLLECTIVE_FALLBACK
is silent and will make the CVAR selecting composition algo work.

* Non compositional algorithms are safe to be selected by CVAR since it
will never create recursive situation. Some "compositional" algorithms
that does not recurse into the same coll_type also fit into this
category.

* For benchmark tests, user may want to set MPIR_CVAR_COLLECTIVE_FALLBACK=error
so they get confirmed to be testing the correct algorithms.

* It is possible for developer to write an algorithm that doesn't have
good restriction check that may end up in a recursive situation, for
now, this is a bug with fault on developers. TODO: add a mechanism to
detect infinite recurse loop.