ch4/ofi: refactor MPIDI_OFI_rdmaread_poll
Refactor the asynchronous pipelined RDMA read operation to be reused by
other parts of the code.

It is a compromise for now to use an MPIR_Request for the async context.
This minimizes the amount of code changes to ofi_rndv_read.c. We do not
need the full MPIR_Request for the RDMA read operation on its own.