build: build single abi with --enable-mpi-abi
Instead of build both mpich abi and mpi abi, only build a single library
with --enable-mpi-abi.

If configured with --enable-mpi-abi, we only install libmpi_abi.so and
mpi.h will be a link to mpi_abi.h. Compiler scripts such as mpicc no
longer supports -mpi-abi option. It will either link to libmpi.so or
libmpi_abi.so based on the configure option.